\name{rrgraph-package}
\alias{rrgraph-package}
\alias{rrgraph}
\docType{package}
\title{
\packageTitle{rrgraph}
}
\description{
\packageDescription{rrgraph}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rrgraph}
\packageIndices{rrgraph}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{rrgraph}

Maintainer: \packageMaintainer{rrgraph}
}
\references{
~~ Literature or other references for background information ~~
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
