\name{stratifiedRrs}
\alias{stratifiedRrs}
\docType{data}
\title{
Significant Relative Risks in each age/sex stratum
}
\description{
Data frame of significant RRs in each age/sex subgroup
}
\usage{data("stratifiedRrs")}
\format{
data.frame
}
\details{
Significant relative risk relations were selected as for \code{\link{defaultGraph}}, 
except that an additional selection criteria was applied -
there must be at more than 7 in each contingency bin.
This criteria was added because relative risk is inaccurate with small numbers the contingency table.
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
}
\keyword{datasets}
